+++ scripts/format.sh
@@ -2,6 +2,6 @@
 # run various linters
 set -e
 echo "running black..."
-python -m black .
+python -m ruff format .
 echo "sorting import with ruff..."
-python -m ruff . --select I,F401 --fix --show-fixes
+python -m ruff check . --select I,F401 --fix --show-fixes