+++ tests/blackbox/test_statsd.py
@@ -156,7 +156,8 @@ def test_from_container(docker_client):
         container.start()
         with removing(container):
             container.wait()
-        sleep(0.1)
+            input("READY!")
+            sleep(0.1)
     assert capture.count("mymet").total() == 1